+++ builder_pkgs/build_modules/lib/build_modules.dart
@@ -2,11 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/common.dart' show multiRootScheme;
+export 'src/common.dart' show fesWorkerPortPath, multiRootScheme;
 export 'src/ddc_names.dart';
 export 'src/errors.dart' show MissingModulesException, UnsupportedModules;
-export 'src/frontend_server_resources.dart'
-    show frontendServerState, frontendServerStateResource;
 export 'src/kernel_builder.dart' show KernelBuilder, reportUnusedKernelInputs;
 export 'src/meta_module_builder.dart'
     show MetaModuleBuilder, metaModuleExtension;
@@ -24,4 +22,6 @@ export 'src/workers.dart'
         dartdevkDriverResource,
         frontendServerProxyDriverResource,
         maxWorkersPerTask,
-        persistentFrontendServerResource;
+        persistentFrontendServer,
+        persistentFrontendServerResource,
+        startFrontendServerWorker;