+++ builder_pkgs/build_modules/lib/src/common.dart
@@ -13,6 +13,11 @@ const webHotReloadOption = 'web-hot-reload';
 final sdkDir = p.dirname(p.dirname(Platform.resolvedExecutable));
 final packagesFilePath = p.join('.dart_tool', 'package_config.json');
 
+/// The file path containing the Frontend Server worker's port.
+///
+/// Used for expression evaluation.
+final fesWorkerPortPath = p.join('.dart_tool', 'build', 'fes_worker_port');
+
 final defaultAnalysisOptionsId = AssetId(
   'build_modules',
   'lib/src/analysis_options.default.yaml',